39-79 рефакторинг кода
