--- packages/server/Theme/readMe.md
@@ -1,12 +1,0 @@
-a. Получение темы пользователя
-	Получение userId из параметров запроса.
-	Поиск в таблице user_theme по ownerId.
-	Получение themeId из найденной записи.
-	Поиск в таблице site_theme по themeId.
-	Возвращение найденной темы(theme) или сообщение об ошибке.
-b. Создание или обновление темы
-	Получение userId, theme и description из запроса.
-	Проверка наличия темы в таблице site_theme.
-	Если тема существует, обновите связь в таблице user_theme.
-	Если тема не существует, создайте новую тему в site_theme и установите связь в user_theme.
-	Возвращение созданной или обновленной темы.
