+++ packages/server/controllers/index.ts
@@ -1,0 +1,1 @@
+export * as themeController from './theme.controller'
