+++ packages/server/db.ts
@@ -1,5 +1,5 @@
-import SiteTheme from './Theme/models/Theme'
-import UserTheme from './Theme/models/User'
+import SiteTheme from './models/theme.model'
+import UserTheme from './models/user.model'
 import 'dotenv/config'
 import {
   Sequelize,
