+++ packages/server/routes/index.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express'
+import themeRouter from './theme.router'
+
+const router = Router()
+
+router.use('', themeRouter)
+
+export default router
